requirements.txt
discord.py
python-dotenv
aiosqlite
jamo
google-genai
plotly 
kaleido 
pandas
captcha
Pillow